cartbox/requirements.txt
psycopg2==2.7.4
Django~=1.11
djangorestframework==3.9.2